requirements.txt
asgiref==3.5.2
attrs==21.4.0
autopep8==1.6.0
click==8.1.3
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
uritemplate==4.1.1
